requirements.txt
playwright==1.49.1
psycopg2-binary==2.9.10
requests==2.32.3
flask==3.0.0
beautifulsoup4==4.12.3
duckduckgo-search==8.1.1
